[Sprint2] API test

+++ controllers/groups/DeleteGroup.js
@@ -3,20 +3,28 @@ const { group, group_user } = require("../../models");
 module.exports = {
   post: async (req, res) => {
     const { groupId, userId } = req.body
-    await group.destroy({
-      where: {
-        id: groupId,
-        admin: userId
-      }
-    })
-    .then(result => {
-      group_user.destroy({
-        where: {
-          group_id: groupId
-        }
+    
+    await group.findOne({ where: { id: groupId, admin: userId}})
+    .then(isAdmin => {
+      if(!isAdmin) {
+        res.status(401).send({ message: "그룹 삭제 권한이 없습니다."})
+      } else {
+        group.destroy({
+          where: {
+            id: groupId,
+            admin: userId
+          }
+        })
+      .then(result => {
+        group_user.destroy({
+          where: {
+            group_id: groupId
+          }
+        })
       })
-    })
-    .then(result => res.status(200).send({ message: "성공적으로 그룹을 삭제했습니다."}))
-    .catch(err => res.status(401).send({ message: "그룹 삭제 권한이 없습니다."}))
+      .then(result => res.status(200).send({ message: "성공적으로 그룹을 삭제했습니다."}))
+      }
+    }
+    )
   }
 };
